improvements
- added controls for failure of functions handling files
    - in agreement with Pan of SpinningKids, used the VS static analyzer in conjunction with /W4 and /Wall to spot potential problems
    - planning to do future investigations using Re-Sharper analyzer

+++ EMULATOR/SRC/funcs.h
@@ -9,9 +9,9 @@ void error(const char * format, ...);
 
 void *alloc(size_t count, size_t size);
 
-void load_buffer(char *file, uint8_t **buffer, uint32_t *size);
+void load_buffer(char *file, uint8_t **buffer, size_t *size);
 
-void save_buffer(char *file, uint8_t *buffer, uint32_t size);
+void save_buffer(char *file, uint8_t *buffer, size_t size);
 
 uint8_t str_to_word(char *str, uint16_t *num);
